refactor alertRule: 서비스 분리
