+++ src/main/java/com/anyang/maruni/domain/alertrule/application/service/AlertDetectionService.java
@@ -11,11 +11,17 @@ import com.anyang.maruni.domain.alertrule.application.config.AlertConfigurationP
 import com.anyang.maruni.domain.alertrule.application.service.AlertAnalysisOrchestrator;
 import com.anyang.maruni.domain.alertrule.domain.entity.AlertRule;
 import com.anyang.maruni.domain.alertrule.domain.entity.AlertType;
+import com.anyang.maruni.domain.alertrule.domain.entity.AlertLevel;
 import com.anyang.maruni.domain.alertrule.domain.repository.AlertRuleRepository;
 import com.anyang.maruni.domain.conversation.domain.entity.MessageEntity;
 import com.anyang.maruni.domain.member.domain.entity.MemberEntity;
+import com.anyang.maruni.domain.alertrule.application.util.AlertServiceUtils;
 
 import lombok.RequiredArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.stream.Collectors;
 
 /**
  * 이상징후 감지 전담 서비스
@@ -31,6 +37,7 @@ public class AlertDetectionService {
     private final AlertRuleRepository alertRuleRepository;
     private final AlertAnalysisOrchestrator analysisOrchestrator;
     private final AlertConfigurationProperties alertConfig;
+    private final AlertServiceUtils alertServiceUtils;
 
     /**
      * 회원의 이상징후 종합 감지
@@ -40,8 +47,10 @@ public class AlertDetectionService {
      */
     @Transactional
     public List<AlertResult> detectAnomalies(Long memberId) {
-        // TODO: Phase 2에서 구현 예정
-        throw new UnsupportedOperationException("Phase 2에서 구현 예정");
+        MemberEntity member = alertServiceUtils.validateAndGetMember(memberId);
+        List<AlertRule> activeRules = alertRuleRepository.findActiveRulesWithMemberAndGuardian(memberId);
+
+        return processAlertRules(member, activeRules);
     }
 
     /**
@@ -53,8 +62,19 @@ public class AlertDetectionService {
      */
     @Transactional
     public AlertResult detectKeywordAlert(MessageEntity message, Long memberId) {
-        // TODO: Phase 2에서 구현 예정
-        throw new UnsupportedOperationException("Phase 2에서 구현 예정");
+        MemberEntity member = alertServiceUtils.validateAndGetMember(memberId);
+
+        // Strategy Pattern을 사용한 키워드 분석
+        AnalysisContext context = AnalysisContext.forKeyword(message);
+        AlertResult keywordResult = analysisOrchestrator.analyzeByType(AlertType.KEYWORD_DETECTION, member, context);
+
+        // 키워드가 감지된 경우, 즉시 알림 처리 고려할 수 있음
+        if (keywordResult.isAlert() && keywordResult.getAlertLevel() == AlertLevel.EMERGENCY) {
+            // 긴급 키워드의 경우 즉시 처리 로직 추가 가능
+            // 현재는 단순히 결과만 반환
+        }
+
+        return keywordResult;
     }
 
     /**
@@ -64,8 +84,7 @@ public class AlertDetectionService {
      * @return 활성 알림 규칙 목록 (Member, Guardian 포함)
      */
     public List<AlertRule> getActiveRulesByMemberId(Long memberId) {
-        // TODO: Phase 2에서 구현 예정
-        throw new UnsupportedOperationException("Phase 2에서 구현 예정");
+        return alertRuleRepository.findActiveRulesWithMemberAndGuardian(memberId);
     }
 
     /**
@@ -75,8 +94,13 @@ public class AlertDetectionService {
      * @return 우선순위 정렬된 활성 알림 규칙 목록
      */
     public List<AlertRule> getActiveRulesByMemberIdOrderedByPriority(Long memberId) {
-        // TODO: Phase 2에서 구현 예정
-        throw new UnsupportedOperationException("Phase 2에서 구현 예정");
+        List<AlertRule> rules = alertRuleRepository.findActiveRulesWithMemberAndGuardian(memberId);
+
+        return rules.stream()
+                .sorted(Comparator.comparing(AlertRule::getAlertLevel, AlertLevel.descendingComparator())
+                    .thenComparing(AlertRule::getAlertType)
+                    .thenComparing(AlertRule::getCreatedAt, Comparator.reverseOrder()))
+                .collect(Collectors.toList());
     }
 
     // ========== Private 메서드들 (Phase 2에서 구현) ==========
@@ -85,31 +109,61 @@ public class AlertDetectionService {
      * 활성 알림 규칙들 처리
      */
     private List<AlertResult> processAlertRules(MemberEntity member, List<AlertRule> activeRules) {
-        // TODO: Phase 2에서 기존 AlertRuleService에서 이동 예정
-        throw new UnsupportedOperationException("Phase 2에서 구현 예정");
+        List<AlertResult> detectedAnomalies = new ArrayList<>();
+
+        for (AlertRule rule : activeRules) {
+            AlertResult analysisResult = analyzeByRuleType(member, rule);
+
+            if (isAlertTriggered(analysisResult)) {
+                detectedAnomalies.add(analysisResult);
+            }
+        }
+
+        return detectedAnomalies;
     }
 
     /**
      * 알림 발생 여부 판단
      */
     private boolean isAlertTriggered(AlertResult analysisResult) {
-        // TODO: Phase 2에서 기존 AlertRuleService에서 이동 예정
-        throw new UnsupportedOperationException("Phase 2에서 구현 예정");
+        return analysisResult != null && analysisResult.isAlert();
     }
 
     /**
      * 알림 규칙 타입별 이상징후 분석
      */
     private AlertResult analyzeByRuleType(MemberEntity member, AlertRule rule) {
-        // TODO: Phase 2에서 기존 AlertRuleService에서 이동 예정
-        throw new UnsupportedOperationException("Phase 2에서 구현 예정");
+        AlertType alertType = rule.getAlertType();
+
+        // 키워드 감지는 실시간 처리이므로 종합 분석에서는 제외
+        if (alertType == AlertType.KEYWORD_DETECTION) {
+            return AlertResult.noAlert();
+        }
+
+        // Strategy Pattern을 사용한 분석
+        if (analysisOrchestrator.isSupported(alertType)) {
+            AnalysisContext context = createAnalysisContext(alertType, alertConfig.getAnalysis().getDefaultDays());
+            return analysisOrchestrator.analyzeByType(alertType, member, context);
+        }
+
+        // 지원하지 않는 타입인 경우
+        return AlertResult.noAlert();
     }
 
     /**
      * 알림 타입에 맞는 분석 컨텍스트 생성
      */
     private AnalysisContext createAnalysisContext(AlertType alertType, int defaultDays) {
-        // TODO: Phase 2에서 기존 AlertRuleService에서 이동 예정
-        throw new UnsupportedOperationException("Phase 2에서 구현 예정");
+        switch (alertType) {
+            case EMOTION_PATTERN:
+                return AnalysisContext.forEmotionPattern(defaultDays);
+            case NO_RESPONSE:
+                return AnalysisContext.forNoResponse(defaultDays);
+            case KEYWORD_DETECTION:
+                // 키워드 분석은 실시간 메시지가 필요하므로 여기서는 빈 컨텍스트
+                return AnalysisContext.builder().build();
+            default:
+                return AnalysisContext.builder().analysisDays(defaultDays).build();
+        }
     }
 }
